added models,py
* this file is going to contain all the models for the project
